init test_rules: break dummy tests for now
