+++ tests/utils.py
@@ -7,3 +7,18 @@ def capture_stdouterr(command, cwd = here):
     p_err = p[1].decode('utf-8').strip()
     return (p_out, p_err)
 
+class TempDirectory(object):
+    def __init__(self):
+        self.tempdir = tempfile.mkdtemp(prefix='sourmashtest_')
+
+    def __enter__(self):
+        return self.tempdir
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        try:
+            shutil.rmtree(self.tempdir, ignore_errors=True)
+        except OSError:
+            pass
+
+        if exc_type:
+            return False
